% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akagenealogy}
\alias{akagenealogy}
\title{akagenealogy}
\format{
A data frame with 92 rows and 13 variables:
\describe{
  \item{\code{egoid}}{double Ego ID}
  \item{\code{age}}{double Ego age}
  \item{\code{sex}}{character Ego sex}
  \item{\code{motherid}}{character Mother ID}
  \item{\code{fatherid}}{character Father ID}
  \item{\code{living}}{character Is ego living?}
  \item{\code{trail}}{character Ego residence trail}
  \item{\code{num_young_children}}{double Number of young children}
  \item{\code{num_older_boys}}{double Number of older boys}
  \item{\code{num_older_girls}}{double Number of older girls}
  \item{\code{num_adult_offspring}}{integer Number of adult offspring}
  \item{\code{total_offspring}}{double Total offspring}
  \item{\code{avg_relatedness}}{double Average relatedness to other members of both trails}
}
}
\usage{
akagenealogy
}
\description{
Genealogy of Aka members of two trails
}
\details{
DETAILS
}
\keyword{datasets}
