% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mturk2011}
\alias{mturk2011}
\title{mturk2011}
\format{
A data frame with 120 rows and 21 variables:
\describe{
  \item{\code{id}}{character MTurk ID}
  \item{\code{seconds}}{double Time to complete the study in seconds}
  \item{\code{consented}}{character Did participant consent to participate?}
  \item{\code{age}}{double Participant age in years}
  \item{\code{sex}}{character Participant sex}
  \item{\code{nation_current}}{character In what country are you living now?}
  \item{\code{nation_home}}{character What country do you consider to be your "home" country?}
  \item{\code{occupation}}{character What is your occupation?}
  \item{\code{office_experience}}{character Have you ever worked in an office?}
  \item{\code{competition}}{double How competitive do you feel with Dave/Elizabeth?}
  \item{\code{condition}}{character Family or work condition}
  \item{\code{work_gossip}}{double Work gossip score (number of positive statements out of 9 total)}
  \item{\code{family_gossip}}{double  (number of positive statements out of 7 total)}
  \item{\code{criticize}}{double How likely is Dave/Elizabeth to criticize you behind your back?}
  \item{\code{family_work}}{double In your experience, how much does work performance/personal affairs influence personal affairs/work performance?}
  \item{\code{friendly}}{double How friendly is Dave/Elizabeth?}
  \item{\code{get_your_job}}{double If Dave/Elizabeth wanted your job as supervisor in this office/Executor, how likely do you think s/he would be to get it?}
  \item{\code{impression}}{double What is your impression of Dave/Elizabeth?}
  \item{\code{interference}}{double If Dave/Elizabeth got into a dispute with a man/woman at work/cousin, how much should the man worry that Dave/Elizabeth would try to interfere with his/her opportunities at work/relationships with other family members?}
  \item{\code{afraid}}{double How physically threatening is Dave/Elizabeth?}
  \item{\code{give_benefit}}{double How likely would you be to give the raise/painting to Dave/Elizabeth?}
}
}
\usage{
mturk2011
}
\description{
Experimental vignette study of gossip, reputation, and resource transfer on Mechanical Turk, 2011
}
\details{
Participants were randomly assigned to the work or family condition, and to a specific distribution of positive and negative work gossip statements and family gossip statements. Data collection started on July 29, 2011, 9:58 am. All but one participant finished by 10:13 am.
}
\keyword{datasets}
