% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngandu2012}
\alias{ngandu2012}
\title{ngandu2012}
\format{
A data frame with 160 rows and 17 variables:
\describe{
  \item{\code{id}}{integer Participant ID code}
  \item{\code{sex}}{character Participant sex}
  \item{\code{age}}{double Participant age (years)}
  \item{\code{language}}{character Language of interview (e.g., Sango, Ngandu, French)}
  \item{\code{condition}}{character Experimental condition (Work/Family)}
  \item{\code{family_gossip}}{double Family gossip score (0=all negative, 7=all positive)}
  \item{\code{work_gossip}}{double Work gossip score (0=all negative, 7=all positive)}
  \item{\code{give_benefit}}{double How likely are you to give the clothing? (unlikely=1, likely=5)}
  \item{\code{impression}}{double What is your opinion of the gossip target? (very bad=1, very good=5)}
  \item{\code{friendly}}{double How friendly is the gossip target? (very unfriendly=1, very friendly=5)}
  \item{\code{afraid}}{double How afraid are you of the gossip target (very unafraid=1, very afraid=5)}
  \item{\code{argue}}{double If you do not give the clothes, how likely would the gossip target argue with you? (unlikely=1, likely=5)}
  \item{\code{criticize}}{double If you do not give the clothing, will the gossip target say bad things about you? (unlikely=1, likely=5)}
  \item{\code{believe_positive}}{double How much do you believe the positive gossip? (don't believe=1, believe=5)}
  \item{\code{believe_negative}}{double How much do you believe the negative gossip? (don't believe=1, believe=5)}
  \item{\code{version}}{character Version of the questionnaire}
  \item{\code{notes}}{character Interview notes}
}
}
\usage{
ngandu2012
}
\description{
Experimental vignette study of gossip, reputation, and resource transfer among Ngandu horticulturalists of the Central African Republic, 2012
}
\details{
Data were collected Summer 2012. Participants were randomly assigned to the work or family condition, and to a specific distribution of positive and negative work gossip statements and family gossip statements.
}
\keyword{datasets}
