% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngandu_ratings}
\alias{ngandu_ratings}
\title{ngandu_ratings}
\format{
A data frame with 537 rows and 9 variables:
\describe{
  \item{\code{seq_id}}{double COLUMN_DESCRIPTION}
  \item{\code{opp}}{character COLUMN_DESCRIPTION}
  \item{\code{condition}}{character Family or work condition}
  \item{\code{sex_informant}}{character Sex of the informant}
  \item{\code{gossip_statement}}{double Code number of gossip statement}
  \item{\code{rating}}{double Informant rating of gossip statement (very bad, bad, neutral, good, very good)}
  \item{\code{gossip_type}}{character Family or work gossip}
  \item{\code{gossip_valence}}{character Positive or negative gossip}
  \item{\code{gossip_text}}{character The text of the gossip statement}
}
}
\usage{
ngandu_ratings
}
\description{
Ratings of each of each gossip statement in family and work conditions
}
\details{
DETAILS
}
\keyword{datasets}
