% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aka_ratings}
\alias{aka_ratings}
\title{aka_ratings}
\format{
A data frame with 40 rows and 9 variables:
\describe{
  \item{\code{id}}{character Target ID}
  \item{\code{reputation}}{double Target reputation}
  \item{\code{rep_rater}}{double Reputation of rater (all targets were also raters)}
  \item{\code{give}}{double How much does target share with others?}
  \item{\code{receive}}{double How much does target receive from others?}
  \item{\code{workhard}}{double How hard does the target work?}
  \item{\code{parenting}}{double How good a parent is the target?}
  \item{\code{angry}}{double How angry is the target?}
}
}
\usage{
aka_ratings
}
\description{
Peer ratings of Aka members of two trails
}
\details{
DETAILS
}
\keyword{datasets}
